// types/cms.tsx
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
  DateTime: any;
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: any;
  /** The `Upload` scalar type represents a file upload. */
  Upload: any;
};

export type Animal = {
  __typename?: 'Animal';
  animal_type?: Maybe<AnimalTypeEntityResponse>;
  createdAt?: Maybe<Scalars['DateTime']>;
  images?: Maybe<UploadFileRelationResponseCollection>;
  leg_count: Scalars['Int'];
  name: Scalars['String'];
  publishedAt?: Maybe<Scalars['DateTime']>;
  updatedAt?: Maybe<Scalars['DateTime']>;
};


export type AnimalImagesArgs = {
  filters?: InputMaybe<UploadFileFiltersInput>;
  pagination?: InputMaybe<PaginationArg>;
  sort?: InputMaybe<Array<InputMaybe<Scalars['String']>>>;
};

export type AnimalEntity = {
  __typename?: 'AnimalEntity';
  attributes?: Maybe<Animal>;
  id?: Maybe<Scalars['ID']>;
};

export type AnimalEntityResponse = {
  __typename?: 'AnimalEntityResponse';
  data?: Maybe<AnimalEntity>;
};

export type AnimalEntityResponseCollection = {
  __typename?: 'AnimalEntityResponseCollection';
  data: Array<AnimalEntity>;
  meta: ResponseCollectionMeta;
};
// ...
export type AnimalRelationResponseCollection = {
  __typename?: 'AnimalRelationResponseCollection';
  data: Array<AnimalEntity>;
};

export type AnimalType = {
  __typename?: 'AnimalType';
  animals?: Maybe<AnimalRelationResponseCollection>;
  createdAt?: Maybe<Scalars['DateTime']>;
  name: Scalars['String'];
  publishedAt?: Maybe<Scalars['DateTime']>;
  updatedAt?: Maybe<Scalars['DateTime']>;
};


export type AnimalTypeAnimalsArgs = {
  filters?: InputMaybe<AnimalFiltersInput>;
  pagination?: InputMaybe<PaginationArg>;
  publicationState?: InputMaybe<PublicationState>;
  sort?: InputMaybe<Array<InputMaybe<Scalars['String']>>>;
};

export type AnimalTypeEntity = {
  __typename?: 'AnimalTypeEntity';
  attributes?: Maybe<AnimalType>;
  id?: Maybe<Scalars['ID']>;
};

export type AnimalTypeEntityResponse = {
  __typename?: 'AnimalTypeEntityResponse';
  data?: Maybe<AnimalTypeEntity>;
};

export type AnimalTypeEntityResponseCollection = {
  __typename?: 'AnimalTypeEntityResponseCollection';
  data: Array<AnimalTypeEntity>;
  meta: ResponseCollectionMeta;
};
// ...
export type I18NLocaleEntityResponse = {
  __typename?: 'I18NLocaleEntityResponse';
  data?: Maybe<I18NLocaleEntity>;
};

export type I18NLocaleEntityResponseCollection = {
  __typename?: 'I18NLocaleEntityResponseCollection';
  data: Array<I18NLocaleEntity>;
  meta: ResponseCollectionMeta;
};
// ...
export type Query = {
  __typename?: 'Query';
  animal?: Maybe<AnimalEntityResponse>;
  animalType?: Maybe<AnimalTypeEntityResponse>;
  animalTypes?: Maybe<AnimalTypeEntityResponseCollection>;
  animals?: Maybe<AnimalEntityResponseCollection>;
  i18NLocale?: Maybe<I18NLocaleEntityResponse>;
  i18NLocales?: Maybe<I18NLocaleEntityResponseCollection>;
  me?: Maybe<UsersPermissionsMe>;
  uploadFile?: Maybe<UploadFileEntityResponse>;
  uploadFiles?: Maybe<UploadFileEntityResponseCollection>;
  usersPermissionsRole?: Maybe<UsersPermissionsRoleEntityResponse>;
  usersPermissionsRoles?: Maybe<UsersPermissionsRoleEntityResponseCollection>;
  usersPermissionsUser?: Maybe<UsersPermissionsUserEntityResponse>;
  usersPermissionsUsers?: Maybe<UsersPermissionsUserEntityResponseCollection>;
};
// ...
export type ResponseCollectionMeta = {
  __typename?: 'ResponseCollectionMeta';
  pagination: Pagination;
};
// ...
export type UploadFileEntityResponse = {
  __typename?: 'UploadFileEntityResponse';
  data?: Maybe<UploadFileEntity>;
};

export type UploadFileEntityResponseCollection = {
  __typename?: 'UploadFileEntityResponseCollection';
  data: Array<UploadFileEntity>;
  meta: ResponseCollectionMeta;
};
// ...
export type UploadFileRelationResponseCollection = {
  __typename?: 'UploadFileRelationResponseCollection';
  data: Array<UploadFileEntity>;
};
// ...
export type UsersPermissionsMe = {
  __typename?: 'UsersPermissionsMe';
  blocked?: Maybe<Scalars['Boolean']>;
  confirmed?: Maybe<Scalars['Boolean']>;
  email?: Maybe<Scalars['String']>;
  id: Scalars['ID'];
  role?: Maybe<UsersPermissionsMeRole>;
  username: Scalars['String'];
};
// ...
export type UsersPermissionsRoleEntityResponse = {
  __typename?: 'UsersPermissionsRoleEntityResponse';
  data?: Maybe<UsersPermissionsRoleEntity>;
};

export type UsersPermissionsRoleEntityResponseCollection = {
  __typename?: 'UsersPermissionsRoleEntityResponseCollection';
  data: Array<UsersPermissionsRoleEntity>;
  meta: ResponseCollectionMeta;
};
// ...
export type UsersPermissionsUserEntityResponse = {
  __typename?: 'UsersPermissionsUserEntityResponse';
  data?: Maybe<UsersPermissionsUserEntity>;
};

export type UsersPermissionsUserEntityResponseCollection = {
  __typename?: 'UsersPermissionsUserEntityResponseCollection';
  data: Array<UsersPermissionsUserEntity>;
  meta: ResponseCollectionMeta;
};
// ...
export type AnimalsByTypeQuery = { __typename?: 'Query', animalTypes?: { __typename?: 'AnimalTypeEntityResponseCollection', data: Array<{ __typename?: 'AnimalTypeEntity', id?: string | null, attributes?: { __typename?: 'AnimalType', name: string, animals?: { __typename?: 'AnimalRelationResponseCollection', data: Array<{ __typename?: 'AnimalEntity', id?: string | null, attributes?: { __typename?: 'Animal', name: string } | null }> } | null } | null }> } | null };
// ...
export const AnimalsByTypeDocument = gql`
    query AnimalsByType {
  animalTypes {
    data {
      id
      attributes {
        animals {
          data {
            id
            attributes {
              name
            }
          }
        }
        name
      }
    }
  }
}
    `;